Major changes to parser and new features

+++ src/main.cpp
@@ -29,20 +29,38 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #define CPPHTTPLIB_OPENSSL_SUPPORT
-#include <iostream>
-#include <fstream>
-#include <filesystem>
-#include <thread>
-#include <httplib.h>
+#include "omtl/ParseTree.hpp"
+#include "omtl/Tokenizer.hpp"
 #include <bxzstr.hpp>
+#include <deb-downloader.hpp>
+#include <estd/ptr.hpp>
+#include <filesystem>
+#include <fstream>
+#include <httplib.h>
+#include <iostream>
+#include <memory>
 #include <tar/tar.hpp>
-#include "Tokenizer.hpp"
+#include <thread>
 
 using namespace std;
 using namespace httplib;
+using namespace omtl;
+using namespace estd::shortnames;
 namespace fs = std::filesystem;
 
-tuple<string,string,string> splitUrl(string url){
+cptr<deb::Installer> debInstaller;
+
+std::string generateUniqueTempDir() {
+    while (true) {
+        std::filesystem::path name = "." + estd::string_util::gen_random(10);
+        if (!std::filesystem::exists(name)) {
+            std::filesystem::create_directories(name);
+            return name.string();
+        }
+    }
+}
+
+tuple<string, string, string> splitUrl(string url) {
     std::string delimiter = " ";
     std::string scheme = "";
     std::string host = "";
@@ -67,10 +85,10 @@ tuple<string,string,string> splitUrl(string url){
 
     path = url;
 
-    return make_tuple(scheme,host,path);
+    return make_tuple(scheme, host, path);
 }
 
-std::filesystem::path downloadFile(string url, std::filesystem::path location){
+std::filesystem::path downloadFile(string url, std::filesystem::path location) {
     std::string scheme = "";
     std::string host = "";
     std::string path = "";
@@ -81,75 +99,59 @@ std::filesystem::path downloadFile(string url, std::filesystem::path location){
     std::filesystem::path filename = extractFilename.filename();
 
     fs::create_directories(location);
-    
-    ofstream file(location/filename);
 
-    httplib::Client cli((scheme+host).c_str());
+    ofstream file(location / filename);
+
+    httplib::Client cli((scheme + host).c_str());
     cli.set_follow_location(true);
 
-    auto res = cli.Get(path.c_str(), Headers(),
-    [&](const Response &response) {
-        return true; // return 'false' if you want to cancel the request.
-    },
-    [&](const char *data, size_t data_length) {
-        file.write(data, data_length);
-        return true; // return 'false' if you want to cancel the request.
-    });
+    auto res = cli.Get(
+        path.c_str(),
+        Headers(),
+        [&](const Response& response) {
+            return true; // return 'false' if you want to cancel the request.
+        },
+        [&](const char* data, size_t data_length) {
+            file.write(data, data_length);
+            return true; // return 'false' if you want to cancel the request.
+        }
+    );
 
     file.close();
-    return location/filename;
+    return location / filename;
 }
 
-std::vector<std::vector<Token>> splitTokens(std::vector<Token> tokens, string delimiter){
-    std::vector<std::vector<Token>> result;
-    std::vector<Token> current;
-    for(std::size_t i = 0; i < tokens.size(); i++){
-        if(tokens[i].value == delimiter){
-            result.push_back(current);
-            current.clear();
-        }else{
-            current.push_back(tokens[i]);
-        }
-    }
-    result.push_back(current);
-    current.clear();
-    return result;
-}
-
-void parseGit(std::vector<Token> tokens, std::filesystem::path root){
-    if(tokens.size() != 5){
-        cerr << "[WARNING] not enough arguments for git statement at " + tokens[0].location << endl;
+void parseGit(Element tokens, std::filesystem::path root) {
+    if (tokens.size() != 5) {
+        cerr << "[WARNING] not enough arguments for git statement at " + tokens.location << endl;
     }
 
-    string sourceType = tokens[0].getString();
-    string sourceUrl = tokens[1].getString();
-    string sourceHash = tokens[2].getString();
-    string source = tokens[3].getString();
-    string destination = tokens[4].getString();
-
+    string sourceUrl = tokens[1]->isString() ? tokens[1]->getString() : tokens[1]->getName();
+    string sourceHash = tokens[2]->isString() ? tokens[2]->getString() : tokens[2]->getName();
+    string source = tokens[3]->isString() ? tokens[3]->getString() : tokens[3]->getName();
+    string destination = tokens[4]->isString() ? tokens[4]->getString() : tokens[4]->getName();
 
     string gitCall = "git clone '";
     gitCall += sourceUrl;
-    gitCall+= "' ";
-    gitCall+= "-b ";
-    gitCall+= sourceHash;
-    gitCall+= " '";;
-    gitCall+=(root / "tmp").string();
-    gitCall+= "'";
+    gitCall += "' ";
+    gitCall += "-b ";
+    gitCall += sourceHash;
+    gitCall += " '";
+
+    gitCall += (root / "tmp").string();
+    gitCall += "'";
 
     cout << gitCall << endl;
-    if(system(gitCall.c_str()) != 0){
-        cout << "git clone for " << sourceUrl << " returned a non zero exit code\n";
-    }
+    if (system(gitCall.c_str()) != 0) { cout << "git clone for " << sourceUrl << " returned a non zero exit code\n"; }
 
-    const auto src = root/"tmp"/source;
-    const auto target = root/destination;
+    const auto src = root / "tmp" / source;
+    const auto target = root / destination;
 
     cout << src.c_str() << endl << target.c_str() << endl;
-    
-    if(fs::is_directory(source)){
+
+    if (fs::is_directory(source)) {
         fs::create_directories(target);
-    }else{
+    } else {
         fs::create_directories(target.parent_path());
     }
 
@@ -158,15 +160,14 @@ void parseGit(std::vector<Token> tokens, std::filesystem::path root){
     fs::remove_all(root / "tmp");
 }
 
-void parseTar(std::vector<Token> tokens, std::filesystem::path root){
-    if(tokens.size() != 4){
-        cerr << "[WARNING] not enough arguments for http statement at " + tokens[0].location << endl;
+void parseTar(Element tokens, std::filesystem::path root) {
+    if (tokens.size() != 4) {
+        cerr << "[WARNING] not enough arguments for tar statement at " + tokens.location << endl;
     }
 
-    string sourceType = tokens[0].getString();
-    string sourceUrl = tokens[1].getString();
-    string source = tokens[2].getString();
-    string destination = tokens[3].getString();
+    string sourceUrl = tokens[1]->isString() ? tokens[1]->getString() : tokens[1]->getName();
+    string source = tokens[2]->isString() ? tokens[2]->getString() : tokens[2]->getName();
+    string destination = tokens[3]->isString() ? tokens[3]->getString() : tokens[3]->getName();
 
     cout << sourceUrl << endl;
 
@@ -175,11 +176,10 @@ void parseTar(std::vector<Token> tokens, std::filesystem::path root){
     bxz::ifstream zFile = bxz::ifstream(filename);
     tar::Reader r(zFile);
 
-    
-    const auto target = root/destination;
+    const auto target = root / destination;
 
     cout << source.c_str() << endl << target.c_str() << endl;
-    
+
     r.extractPath(source, target);
 
     cout << endl;
@@ -187,21 +187,70 @@ void parseTar(std::vector<Token> tokens, std::filesystem::path root){
     fs::remove_all(root / "tmp");
 }
 
-int main(){
-    const auto root = fs::current_path();
-    
-    Tokenizer tkn;
-    ifstream configFile("vendor.txt");
-    auto statements = splitTokens(tkn.tokenize(configFile), ",");
+void parseDebInit(Element tokens, std::filesystem::path root) {
+    std::vector<std::string> sources;
+    if (tokens.size() < 2) {
+        cerr << "[WARNING] not enough arguments for deb-repo statement at " + tokens.location << endl;
+    }
+    if (tokens[1]->isTuple()) {
+        for (size_t i = 0; i < tokens[1]->size(); i++) {
+            auto line = tokens[1][i];
+            if (!line->isString()) {
+                cerr << "[WARNING] debian repository must be a string " + line->location << endl;
+                continue;
+            }
+            cerr << "Added: " << line->getString() << endl;
+            sources.push_back(line->getString());
+        }
+        debInstaller = new deb::Installer(sources);
+    } else {
+        cerr << "[WARNING] bad arguments for deb-repo statement at " + tokens.location << endl;
+    }
+}
 
+void parseDebMarkInstall(Element tokens, std::filesystem::path root) {
+    if (tokens.size() < 2) {
+        cerr << "[WARNING] not enough arguments for deb-ignore statement at " + tokens.location << endl;
+    }
+    for (size_t i = 1; i < tokens.size(); i++) { debInstaller->markPreInstalled({tokens[i]->getString()}); }
+}
+
+void parseDebInstall(Element tokens, std::filesystem::path root) {
+    if (tokens.size() != 4) {
+        cerr << "[WARNING] not enough arguments for deb statement at " + tokens.location << endl;
+    }
+    debInstaller->install(
+        tokens[1]->isString() ? tokens[1]->getString() : tokens[1]->getName(),
+        {
+            {tokens[2]->getString(), tokens[3]->getString()},
+        }
+    );
+    debInstaller->clearInstalled();
+}
+
+void parseInclude(Element cmd, std::filesystem::path root) {
+    if (cmd.size() != 2) throw runtime_error("invalid include command at " + cmd.location);
+    ifstream configFile(cmd[1]->isString() ? cmd[1]->getString() : cmd[1]->getName());
+    Tokenizer tkn;
+    ParseTreeBuilder ptb;
+
+    auto pt = ptb.buildParseTree(tkn.tokenize(configFile));
 
     fs::remove_all(root / "tmp");
-    for(auto tokens: statements){
-        if(tokens.size() <= 0) continue;
-        if(tokens[0].getString() == "git")
-            parseGit(tokens, root);
-        if(tokens[0].getString() == "tar")
-            parseTar(tokens, root);
+
+    for (size_t i = 0; i < pt.size(); i++) {
+        if (pt[i]->size() <= 0) continue;
+        if (!pt[i][0]->isName()) continue;
+        if (pt[i][0]->getName() == "git") parseGit(pt[i].value(), root);
+        if (pt[i][0]->getName() == "tar") parseTar(pt[i].value(), root);
+        if (pt[i][0]->getName() == "deb-init") parseDebInit(pt[i].value(), root);
+        if (pt[i][0]->getName() == "deb-ignore") parseDebMarkInstall(pt[i].value(), root);
+        if (pt[i][0]->getName() == "deb") parseDebInstall(pt[i].value(), root);
+        if (pt[i][0]->getName() == "include") parseInclude(pt[i].value(), root);
     }
+}
+
+int main() {
+    parseInclude(Element({Token("include"), Token("vendor.txt")}), fs::current_path());
     return 0;
 }
